error handler

+++ tests/actions.py
@@ -1,9 +1,10 @@
 from flask import jsonify, request
 from app import app, db
 from app.models import User, Post, Comment
+from .test_data.fills import create_users, create_comments, create_posts
 
 
-@app.route('/test/comments', methods=['GET', 'DELETE'])
+@app.route('/test/comments', methods=['GET', 'DELETE', 'POST'])
 def _comments():
     if request.method == 'GET':
         comments = Comment.query.all()
@@ -18,15 +19,28 @@ def _comments():
                 } for comment in comments
             ]
         })
-    else:
+    elif request.method == 'DELETE':
         comments = Comment.query.all()
         for comment in comments:
             db.session.delete(comment)
         db.session.commit()
         return jsonify({'status': 'deleted'})
+    else:
+        comments = create_comments(2)
+        return jsonify({
+            'comments': [
+                {
+                    'id': comment.id,
+                    'body': comment.body[0:20],
+                    'timestamp': comment.timestamp,
+                    'user_id': comment.user_id,
+                    'post_id': comment.post_id
+                } for comment in comments
+            ]
+        })
 
 
-@app.route('/test/posts', methods=['GET', 'DELETE'])
+@app.route('/test/posts', methods=['GET', 'DELETE', 'POST'])
 def _posts():
     if request.method == 'GET':
         posts = Post.query.all()
@@ -40,15 +54,27 @@ def _posts():
                 } for post in posts
             ]
         })
-    else:
+    elif request.method == 'DELETE':
         posts = Post.query.all()
         for post in posts:
             db.session.delete(post)
         db.session.commit()
         return jsonify({'status': 'deleted'})
+    else:
+        posts = create_posts(2)
+        return jsonify({
+            'posts': [
+                {
+                    'id': post.id,
+                    'body': post.body[0:20],
+                    'timestamp': post.timestamp,
+                    'user_id': post.user_id
+                } for post in posts
+            ]
+        })
 
 
-@app.route('/test/users', methods=['GET', 'DELETE'])
+@app.route('/test/users', methods=['GET', 'DELETE', 'POST'])
 def _users():
     if request.method == 'GET':
         users = User.query.all()
@@ -62,9 +88,18 @@ def _users():
                 } for user in users
             ]
         })
-    else:
+    elif request.method == 'DELETE':
         users = User.query.all()
         for user in users:
             db.session.delete(user)
         db.session.commit()
         return jsonify({'status': 'deleted'})
+    else:
+        users = create_users(10)
+        return jsonify({
+            'users': [{
+                'id': user.id,
+                'username': user.username,
+                'email': user.email
+            } for user in users]
+        })
